requirement.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
anthropic>=0.40.0
httpx>=0.27.0
langgraph
langchain-core
rich==13.7.0
python-dotenv==1.0.0